+++ CHANGES.rst
@@ -39,6 +39,7 @@ Unreleased Changes
   * When running under Docker/Gunicorn, append the decimal number of seconds taken to serve the request to the Gunicorn access log.
 
 * `Issue #184 <https://github.com/jantman/biweeklybudget/issues/184>`_ - Redact database password from ``/help`` view, and change ``/help`` view to show Version containing git commit hash for pre-release/development Docker builds.
+* `Issue #183 <https://github.com/jantman/biweeklybudget/issues/183>`_ - Add UI link to ignore reconciling an OFXTransaction if there will not be a matching Transaction.
 
 0.7.1 (2018-01-10)
 ------------------